backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
mangum==0.19.0
boto3==1.35.99
python-geohash==0.8.5
pytest>=8.0.0
moto[dynamodb]==5.0.28
